#1992 - Combined the hooks to modify column data.
